airflow/requirements.txt
apache-airflow==2.10.4
apache-airflow-providers-apache-kafka==1.7.0
ffmpeg-python==0.2.0
confluent-kafka==2.8.0
boto3==1.36.0
smart-open==7.1.0
python-dotenv==1.0.1
torch==2.5.1
torchvision==0.20.1
pillow==11.1.0